+++ almond/core/app.cpp
@@ -18,16 +18,20 @@ void nd::App::setup(const std::string& filename, sf::Vector2f pos, sf::Vector2f 
 
 // -----------------------------------------------------------------------------
 nd::Widget* nd::App::get_widget(std::string id) { // FUNC@get_widget
-    return _manager_drawables.get_widget_by_id(id);
+    return _drawable_man.get_widget_by_id(id);
 } // END@get_widget
 
 
 // -----------------------------------------------------------------------------
 void nd::App::manage_events() { // FUNC@manage_events
-    if (__root == nullptr) return;
-    while (const std::optional event = __window.pollEvent()) {
-        if (_internal_manage_event(event)) continue;
-        __root->handle_event(event);
+    while (const std::optional sf_event = __window.pollEvent()) {
+        nd::Event nd_event = __init_event(sf_event);
+        _event_man.handle_event(nd_event);
+
+        for (nd::Widget* widget : _drawable_man.get_all_widgets()) {
+            bool consumed = widget->handle_event(nd_event);
+            if (consumed) break;
+        }
     }
 } // END@manage_events
 
@@ -42,7 +46,7 @@ void nd::App::draw() { // FUNC@draw
 // -----------------------------------------------------------------------------
 void nd::App::__create(const std::string& filename) { // FUNC@__create
     if (filename != "") {
-        __root = nd::ParserNDG(_manager_drawables).parse(filename);
+        __root = nd::ParserNDG(_drawable_man).parse(filename);
         if (__root == nullptr) {
             std::cerr << "Failed to parse NDG file: " << filename << std::endl;
             std::cout << "Creating empty Container as the root widget." << std::endl;
@@ -52,7 +56,7 @@ void nd::App::__create(const std::string& filename) { // FUNC@__create
         std::cout << "No NDG file provided. Creating empty Container as the root widget." << std::endl;
         __root = new nd::Container();
     }
-    _manager_drawables.set_id(__root, "root");
+    _drawable_man.set_id(__root, "root");
 } // END@__create
 
 
@@ -61,7 +65,7 @@ void nd::App::__build(sf::Vector2f pos, sf::Vector2f size) { // FUNC@__build
     __root->set_pos(pos);
     __root->set_size(size);
     __root->build();
-    _manager_drawables.group_radiobuttons();
+    _drawable_man.group_radiobuttons();
 } // END@__build