+++ almond/widgets/buttons/button_primitive.hpp
@@ -6,6 +6,7 @@ class ButtonPrimitive : public Widget {
 public:
     bool set_spec(std::string key, std::string raw_value) override; // HEAD@set_spec
     void build() override; // HEAD@build
+    bool handle_event(nd::Event event) override; // HEAD@handle_event
 
     void      set_enabled    (bool enabled); // HEAD@set_enabled
     void      set_bg_idle    (sf::Color color) { __bg_idle = color;     }
@@ -18,12 +19,14 @@ public:
     sf::Color get_bg_pressed () { return __bg_pressed;  }
     sf::Color get_bg_disabled() { return __bg_disabled; }
 
+    void link_on_click(CALLBACK_EVENT callback) { _on_click = callback; }
+
 protected:
     ButtonPrimitive() : Widget() { _bg_color = __bg_idle; }
 
-    bool _internal_on_mouse_button_pressed(const std::optional<sf::Event> event) override; // HEAD@_internal_on_mouse_button_pressed
-    bool _internal_on_mouse_button_released(const std::optional<sf::Event> event) override; // HEAD@_internal_on_mouse_button_released
-    bool _internal_on_mouse_moved(const std::optional<sf::Event> event) override; // HEAD@_internal_on_mouse_moved
+    virtual bool _on_mouse_button_pressed(nd::Event event); // HEAD@_on_mouse_button_pressed
+    virtual bool _on_mouse_button_released(nd::Event event); // HEAD@_on_mouse_button_released
+    virtual bool _on_mouse_moved(nd::Event event); // HEAD@_on_mouse_moved
 
     enum class State {
         IDLE,    // button isn't pressed and mouse isn't over it
@@ -31,6 +34,8 @@ protected:
         PRESSED, // button is pressed. Mouse could be over it or not
         DISABLED // button is disabled and can't be interacted with
     } _state = State::IDLE;
+
+    CALLBACK_EVENT _on_click;
 
 private:
     void __set_state(State state); // HEAD@__set_state