+++ almond/widgets/buttons/check_box.hpp
@@ -10,9 +10,6 @@ public:
     void build() override; // HEAD@build
     void draw(sf::RenderWindow& window) override; // HEAD@draw
 
-protected:
-    void _internal_on_toggle() override; // HEAD@_internal_on_toggle
-
 private:
     sf::VertexArray __check_mark = sf::VertexArray(sf::PrimitiveType::TriangleStrip, 6);
 };