+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [0.3.1] - 2026-07-29
-- Added getter/setter for `sf::ManagerDrawables::__font` (useful for creating custom widgets that need fonts).
+- Added getter/setter for `sf::DrawableManager::__font` (useful for creating custom widgets that need fonts).
 
 ## [0.3.0] - 2026-07-28
 - ALMOND now statically links to a global installation of SFML (instead of each almond project building it locally).